Moved data reading into data module
